requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
langchain>=0.1.13
langchain-google-genai>=1.0.1
google-generativeai>=0.4.1
faiss-cpu>=1.8.0
PyPDF2>=3.0.0
firebase-admin>=6.5.0
httpx>=0.27.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
google-genai
streamlit
sentence-transformers>=2.6.1
numpy>=1.24.0
